requirements.txt
# Backend Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
redis>=5.0.1
asyncpg>=0.30.0
psutil>=5.9.6
structlog>=23.2.0
python-json-logger>=3.2.1
pyyaml>=6.0.2

# AI/ML Dependencies
transformers>=4.38.1
scikit-learn>=1.3.2
numpy>=1.26.2
pandas>=2.1.3

# Database
sqlalchemy>=2.0.23
pymongo>=4.6.1
psycopg2-binary>=2.9.9

# Testing
pytest>=7.4.3
pytest-cov>=6.0.0
pytest-asyncio>=0.25.2

# Monitoring
sentry-sdk>=1.40.4
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
prometheus-client>=0.19.0
astor>=0.8.1
